backend/pyproject.toml
[tool.poetry]
name = "water-portal-api"
version = "0.1.0"
description = ""
authors = ["ymarcon <yannick.marcon@epfl.ch>"]
readme = "README.md"
packages = [{include = "api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
pyproj = "^3.6.1"
requests = "^2.32.3"
pydantic-settings = "^2.4.0"
dynaconf = "^3.2.6"
pillow = "^10.4.0"
aiobotocore = "^2.13.3"
timezonefinder = "^6.5.2"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
python-keycloak = "^4.6.2"
fastapi-cache2 = "^0.2.2"
redis = "^5.1.1"
influxdb-client = "^1.47.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
